requirements.txt
absl-py==0.7.1
astor==0.8.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
dill==0.3.0
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
googleapis-common-protos==1.6.0
grpcio==1.23.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
imgaug==0.2.9
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
networkx==2.3
numpy==1.16.5
opencv-python==4.1.0.25
opt-einsum==3.1.0
Pillow==6.1.0
promise==2.2.1
protobuf==3.9.1
psutil==5.6.3
pyparsing==2.4.2
python-dateutil==2.8.0
PyWavelets==1.0.3
requests==2.22.0
scikit-image==0.15.0
scipy==1.3.1
Shapely==1.6.4.post2
six==1.12.0
tb-nightly==1.14.0a20190603
tensorboard==2.0.0
tensorflow-datasets==1.2.0
tensorflow-estimator==2.0.0
tensorflow-examples===9fb96a113917625db88e306218368fa6a0b807e9-
tensorflow-gpu==2.0.0
tensorflow-metadata==0.14.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
tqdm==4.34.0
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2